+++ src/main/java/insane96mcp/survivalreimagined/module/experience/enchantments/EnchantmentsFeature.java
@@ -66,7 +66,7 @@ public class EnchantmentsFeature extends SRFeature {
 	public static final RegistryObject<Enchantment> MA_JUMP = SRRegistries.ENCHANTMENTS.register("ma_jump", DoubleJump::new);
 	public static final RegistryObject<Enchantment> GRAVITY_DEFYING = SRRegistries.ENCHANTMENTS.register("gravity_defying", GravityDefying::new);
 	public static final RegistryObject<Enchantment> CRITICAL = SRRegistries.ENCHANTMENTS.register("critical", Critical::new);
-	public static final RegistryObject<Enchantment> RHYTHMIC_SWING = SRRegistries.ENCHANTMENTS.register("rhythmic_swing", SleightOfHand::new);
+	public static final RegistryObject<Enchantment> SWIFT_STRIKE = SRRegistries.ENCHANTMENTS.register("swift_strike", SwiftStrike::new);
 	public static final RegistryObject<Enchantment> HEALTHY = SRRegistries.ENCHANTMENTS.register("healthy", Healthy::new);
 
 	@Config
@@ -173,7 +173,7 @@ public class EnchantmentsFeature extends SRFeature {
 		StepUp.applyAttributeModifier(event);
 		GravityDefying.applyAttributeModifier(event);
 		Healthy.applyAttributeModifier(event);
-		SleightOfHand.applyAttributeModifier(event);
+		SwiftStrike.applyAttributeModifier(event);
 	}
 
 	@SubscribeEvent